[front] GetStartedPage: split sub-components into a folder (smb2268)
Structural only, no behavior change: move the page's sub-components into a
GetStartedPage/ folder so index.tsx is just the exported page. Extract
SourceIcon, RecommendationItem, JustAskComposer, RecentConversations,
PreviouslyDoneRow, and the shared recency helpers into their own files.
